refactor: rename ui component file names to pascal case
