+++ apps/web/src/components/ui/WeeklyMvp.tsx
